+++ src/map/ducks/layers/map-layers.js
@@ -1,7 +1,7 @@
 import { createSelector } from 'reselect';
 
 import { createUrlWithToken } from '../../../shared/services/api/api';
-import { getMapOverlays } from '../map/map';
+import { getMapOverlays } from '../map/map-selectors';
 import MAP_CONFIG from '../../services/map-config';
 
 export const FETCH_MAP_LAYERS_REQUEST = 'FETCH_MAP_LAYERS_REQUEST';
